requirements.txt
streamlit
reportlab
pillow
svglib
requests
yt-dlp
